+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Rename `RmContainerOpts` -> `ContainerRemoveOpts`
 - Rename `RmImageOpts` -> `ImageRemoveOpts`
 - Add `Clone` implementation to `ContainerCreateOpts`
+- Add `Clone` implementation to `ImageBuildOpts`
 
 # 0.11.0
 - *BREAKING* Rename `BuildOpts` -> `ImageBuildOpts` and `BuildOptsBuilder` -> `ImageBuildOptsBuilder`